revert removal of metadata chunk grid attribute
